Add files via upload
The assignment was as follows: You are required to write a guide for yourself that explains how to use Python to implement data exploration and cleaning. Your guide should address tasks such as reading a dataset, describing a dataset (fields and size), visualising features of a dataset, performing correlation analysis, handling missing data, handling outliers, and so on.